move yaml to tembo-stacks
