+++ tembo-stacks/src/stacks/specs/gis.yaml
