+++ tembo-stacks/src/stacks/specs/machine_learning.yaml
