+++ tembo-stacks/src/stacks/specs/message_queue.yaml
