+++ tembo-stacks/src/stacks/specs/mongo_alternative.yaml
