+++ tembo-stacks/src/stacks/specs/standard.yaml
