+++ tembo-stacks/src/stacks/specs/vectordb.yaml
